when checking uniq tp when storing, also check uuid
Since it's very plausible that some cats might just have their
phone name iPhone and push a point from the same nanosecond as
another luddite.

I'm pretty sure I've made this change before, but not sure where,
or if I pushed it. Also not sure if UUID logic is in the latest
mobCat app yet.